Fix : 공통 MUI로 수정, 폴더 정리
