+++ public/index.html
@@ -1,32 +1,26 @@
 <!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <!-- 구글 폰트 -->
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700;900&display=swap"
-      rel="stylesheet"
-    />
-    <!-- 카카오 우편번호 조회 -->
-    <script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
-    <script
-      type="text/javascript"
-      src="https://openapi.map.naver.com/openapi/v3/map3.js?ncpClientId=nh1l9cbip1"
-    ></script>
-    <!-- 네이버 로그인 -->
-    <script
-      type="text/javascript"
-      src="https://static.nid.naver.com/js/naveridlogin_js_sdk_2.0.0.js"
-      charset="utf-8"
-    ></script>
-    <title>로스리더</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+<html lang="ko">
+
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="theme-color" content="#000000" />
+  <!-- 구글 폰트 -->
+  <link rel="preconnect" href="https://fonts.googleapis.com" />
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700;900&display=swap" rel="stylesheet" />
+  <!-- 카카오 우편번호 조회 -->
+  <script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
+  <script type="text/javascript" src="https://openapi.map.naver.com/openapi/v3/map3.js?ncpClientId=nh1l9cbip1"></script>
+  <!-- 네이버 로그인 -->
+  <script type="text/javascript" src="https://static.nid.naver.com/js/naveridlogin_js_sdk_2.0.0.js"
+    charset="utf-8"></script>
+  <title>로스리더</title>
+</head>
+
+<body>
+  <div id="root"></div>
+</body>
+
 </html>
