+++ src/Routers.jsx
@@ -44,8 +44,8 @@ const Routers = () => {
         <Route exact path='/signup' component={SignUp} />
         <Route exact path='/product/:productId' component={Detail} />
         <Route exact path='/apply/:productId' component={Apply} />
-        <Route path={'/my'} component={My} />
-        <Route exact path={'/list'} component={List} />
+        <Route path='/my' component={My} />
+        <Route exact path='/list' component={List} />
         <Route path='*' render={() => <Redirect to='/' />} />
       </Switch>
       <Footer />
