+++ src/pages/My/Buy/BuyPage.jsx
@@ -1,6 +1,6 @@
-import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
+import { client } from 'utils/api';
 import Orders from './Orders';
 import UserInfo from './UserInfo';
 
@@ -9,8 +9,12 @@ const BuyPage = () => {
 
   useEffect(() => {
     const fetchOrders = async () => {
-      const { data } = await axios.get(`/data/orders.json`);
-      setOrders(data.data);
+      try {
+        const { data } = await client.get(`/data/orders.json`);
+        setOrders(data.data);
+      } catch (error) {
+        alert(error.message);
+      }
     };
     fetchOrders();
   }, []);
