+++ src/pages/My/Review/ReviewPage.jsx
@@ -1,25 +1,37 @@
-import axios from 'axios';
-import Title from 'Components/common/Title';
-import Reviews from 'pages/My/Review/Reviews';
+import Title from 'Components/Title';
 import React, { useEffect, useState } from 'react';
-
+import styled from 'styled-components';
+import { client } from 'utils/api';
+import Review from './Review';
 const ReviewPage = () => {
   const [reviews, setReviews] = useState([]);
 
   useEffect(() => {
     const fetchReviews = async () => {
-      const { data } = await axios.get('/data/reviews.json');
-      setReviews(data.data);
+      try {
+        const { data } = await client.get('/data/reviews.json');
+        setReviews(data.data);
+      } catch (error) {
+        alert(error.message);
+      }
     };
     fetchReviews();
   }, []);
 
   return (
     <div>
-      <Title text='리뷰관리' m='0 0 0.9375rem 0' />
-      <Reviews reviews={reviews} />
+      <Title text='리뷰관리' margin='0 0 0.9375rem 0' />
+      <ReviewsWrapper>
+        {reviews.map(item => (
+          <Review key={item.id} {...item} />
+        ))}
+      </ReviewsWrapper>
     </div>
   );
 };
+
+const ReviewsWrapper = styled.div`
+  padding-top: 0.625rem;
+`;
 
 export default ReviewPage;
